Move `egui::util::cache` to `egui::cache`, and improve docstrings
