+++ crates/egui/src/util/mod.rs
@@ -1,6 +1,5 @@
 //! Miscellaneous tools used by the rest of egui.
 
-pub mod cache;
 pub(crate) mod fixed_cache;
 pub mod id_type_map;
 pub mod undoer;
@@ -9,3 +8,7 @@ pub use id_type_map::IdTypeMap;
 
 pub use epaint::emath::History;
 pub use epaint::util::{hash, hash_with};
+
+/// Deprecated alias for [`egui::cache`].
+#[deprecated = "Use egui::cache instead"]
+pub use crate::cache;
